+++ common/src/main/java/io/github/flemmli97/mobbattle/client/ClientHandler.java
@@ -3,6 +3,7 @@ package io.github.flemmli97.mobbattle.client;
 import com.mojang.blaze3d.vertex.PoseStack;
 import io.github.flemmli97.mobbattle.MobBattle;
 import io.github.flemmli97.mobbattle.client.gui.EffectScreen;
+import io.github.flemmli97.mobbattle.client.gui.PerimeterSettingScreen;
 import io.github.flemmli97.mobbattle.client.gui.SpawnEggScreen;
 import io.github.flemmli97.mobbattle.common.components.AreaPositionComponent;
 import io.github.flemmli97.mobbattle.common.components.UuidComponent;
@@ -13,6 +14,7 @@ import io.github.flemmli97.mobbattle.common.registry.MobBattleDataComponents;
 import io.github.flemmli97.mobbattle.common.registry.MobBattleItems;
 import io.github.flemmli97.mobbattle.common.utils.Utils;
 import io.github.flemmli97.mobbattle.network.C2SItemFunctionPress;
+import io.github.flemmli97.mobbattle.network.S2CSpawnEggScreen;
 import io.github.flemmli97.mobbattle.platform.CrossPlatformStuff;
 import net.minecraft.client.Camera;
 import net.minecraft.client.KeyMapping;
@@ -30,6 +32,8 @@ import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.EntityHitResult;
 import net.minecraft.world.phys.HitResult;
 import net.minecraft.world.phys.Vec3;
+import net.minecraft.world.phys.shapes.Shapes;
+import net.minecraft.world.phys.shapes.VoxelShape;
 import org.joml.Quaternionf;
 import org.lwjgl.glfw.GLFW;
 
@@ -75,8 +79,9 @@ public class ClientHandler {
         if (heldItem.getItem() == MobBattleItems.MOB_ARMY.get() || heldItem.getItem() == MobBattleItems.MOB_EQUIP.get()) {
             AreaPositionComponent comp = heldItem.get(MobBattleDataComponents.BOX.get());
             if (comp != null && comp.first() != null && comp.second() != null)
-                ClientHandler.renderBlockOutline(stack, Minecraft.getInstance().renderBuffers().crumblingBufferSource(), comp.first(), comp.second());
+                ClientHandler.renderBlockOutline(stack, Minecraft.getInstance().renderBuffers().bufferSource(), comp.first(), comp.second());
         }
+        ClientPerimeterData.render(player, stack);
     }
 
     public static void openEffectGui() {
@@ -84,19 +89,26 @@ public class ClientHandler {
     }
 
     public static void renderBlockOutline(PoseStack stack, MultiBufferSource.BufferSource buffer, BlockPos pos, BlockPos pos2) {
+        AABB aabb = Utils.getBoundingBoxPositions(pos, pos2).deflate(0.05);
+        renderShapeOutline(stack, buffer, Shapes.create(aabb), 0.5F, 1, 1, 0.5f);
+    }
+
+    public static void renderShapeOutline(PoseStack stack, MultiBufferSource.BufferSource buffer, VoxelShape shape, float red, float green, float blue, float alpha) {
         Camera camera = Minecraft.getInstance().gameRenderer.getMainCamera();
+        Vec3 vec = camera.getPosition();
         stack.pushPose();
         stack.mulPose(camera.rotation().conjugate(new Quaternionf()));
-        AABB aabb = Utils.getBoundingBoxPositions(pos, pos2).deflate(0.05);
-        Vec3 vec = camera.getPosition();
         stack.translate(-vec.x, -vec.y, -vec.z);
-        LevelRenderer.renderLineBox(stack, buffer.getBuffer(RenderType.lines()), aabb, 1, 0.5F, 0.5F, 1);
+        LevelRenderer.renderVoxelShape(stack, buffer.getBuffer(RenderType.lines()), shape, 0, 0, 0, red, green, blue, alpha, false);
         buffer.endBatch(RenderType.LINES);
         stack.popPose();
     }
 
-    public static void openSpawneggGui(InteractionHand hand) {
-        Minecraft.getInstance().setScreen(new SpawnEggScreen(hand));
+    public static void openItemScreen(InteractionHand hand, S2CSpawnEggScreen.ScreenType type) {
+        switch (type) {
+            case SPAWN_EGG -> Minecraft.getInstance().setScreen(new SpawnEggScreen(hand));
+            case PERIMETER -> Minecraft.getInstance().setScreen(new PerimeterSettingScreen(hand));
+        }
     }
 
     public static boolean handleEntityHighlight(Entity entity) {