+++ common/src/main/java/io/github/flemmli97/mobbattle/common/items/ItemExtendedSpawnEgg.java
@@ -155,7 +155,7 @@ public class ItemExtendedSpawnEgg extends Item implements ExtendedItem {
                 }
             }
         } else {
-            CrossPlatformStuff.INSTANCE.sendToClient(new S2CSpawnEggScreen(hand), serverPlayer);
+            CrossPlatformStuff.INSTANCE.sendToClient(new S2CSpawnEggScreen(hand, S2CSpawnEggScreen.ScreenType.SPAWN_EGG), serverPlayer);
         }
         return InteractionResultHolder.consume(stack);
     }