+++ fabric/src/main/java/io/github/flemmli97/mobbattle/fabric/client/MobBattleFabricClient.java
@@ -2,6 +2,7 @@ package io.github.flemmli97.mobbattle.fabric.client;
 
 import io.github.flemmli97.mobbattle.client.BossBarItemColor;
 import io.github.flemmli97.mobbattle.client.ClientHandler;
+import io.github.flemmli97.mobbattle.client.ItemModelProps;
 import io.github.flemmli97.mobbattle.client.MultiItemColor;
 import io.github.flemmli97.mobbattle.client.gui.ArmorScreen;
 import io.github.flemmli97.mobbattle.common.registry.MobBattleItems;
@@ -12,6 +13,7 @@ import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.rendering.v1.ColorProviderRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderEvents;
 import net.minecraft.client.gui.screens.MenuScreens;
+import net.minecraft.client.renderer.item.ItemProperties;
 
 public class MobBattleFabricClient implements ClientModInitializer {
 
@@ -23,5 +25,6 @@ public class MobBattleFabricClient implements ClientModInitializer {
         MenuScreens.register(MobBattleMenuTypes.ARMOR_MENU.get(), ArmorScreen::new);
         ClientHandler.registerKeyBinding(KeyBindingHelper::registerKeyBinding);
         ClientTickEvents.END_CLIENT_TICK.register(client -> ClientHandler.keyEvent());
+        ItemProperties.register(MobBattleItems.PERIMETER_TOOL.get(), ItemModelProps.PERIMETER_REMOVE, ItemModelProps.PERIMETER_REMOVE_PROPERTY);
     }
 }