+++ neoforge/src/main/java/io/github/flemmli97/mobbattle/neoforge/client/ClientEvents.java
@@ -2,14 +2,17 @@ package io.github.flemmli97.mobbattle.neoforge.client;
 
 import io.github.flemmli97.mobbattle.client.BossBarItemColor;
 import io.github.flemmli97.mobbattle.client.ClientHandler;
+import io.github.flemmli97.mobbattle.client.ItemModelProps;
 import io.github.flemmli97.mobbattle.client.MultiItemColor;
 import io.github.flemmli97.mobbattle.client.gui.ArmorScreen;
 import io.github.flemmli97.mobbattle.common.registry.MobBattleItems;
 import io.github.flemmli97.mobbattle.common.registry.MobBattleMenuTypes;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.client.event.ClientTickEvent;
 import net.neoforged.neoforge.client.event.RegisterColorHandlersEvent;
 import net.neoforged.neoforge.client.event.RegisterKeyMappingsEvent;
@@ -21,6 +24,7 @@ public class ClientEvents {
 
     public static void register(IEventBus modBus) {
         NeoForge.EVENT_BUS.register(new ClientEvents());
+        modBus.addListener(ClientEvents::clientSetup);
         modBus.addListener(ClientEvents::spawnEggColor);
         modBus.addListener(ClientEvents::menuRegister);
         modBus.addListener(ClientEvents::keyRegister);
@@ -36,6 +40,10 @@ public class ClientEvents {
     @SubscribeEvent(receiveCanceled = true)
     public void keyEvent(ClientTickEvent.Post event) {
         ClientHandler.keyEvent();
+    }
+
+    public static void clientSetup(FMLClientSetupEvent event) {
+        event.enqueueWork(() -> ItemProperties.register(MobBattleItems.PERIMETER_TOOL.get(), ItemModelProps.PERIMETER_REMOVE, ItemModelProps.PERIMETER_REMOVE_PROPERTY));
     }
 
     public static void spawnEggColor(RegisterColorHandlersEvent.Item e) {